# Replace in-python quota calc with a SQL query