--- a/scripts/set_user_disk_usage.py
+++ b/scripts/set_user_disk_usage.py
@@ -41,15 +41,13 @@ def quotacheck(sa_session, users, engine):
     sa_session.refresh(user)
     current = user.get_disk_usage()
     print(user.username, '<' + user.email + '>:', end=' ')
-    if engine not in ('postgres', 'postgresql'):
-        new = user.calculate_disk_usage()
-        sa_session.refresh(user)
-        # usage changed while calculating, do it again
-        if user.get_disk_usage() != current:
-            print('usage changed while calculating, trying again...')
-            return quotacheck(sa_session, user, engine)
+
+    if not args.dryrun:
+        pgcalc(sa_session, user.id, dryrun=False)
+        new = user.get_disk_usage()
     else:
-        new = pgcalc(sa_session, user.id, dryrun=args.dryrun)
+        new = user.calculate_disk_usage()
+
     # yes, still a small race condition between here and the flush
     print('old usage:', nice_size(current), 'change:', end=' ')
     if new in (current, None):
@@ -59,10 +57,6 @@ def quotacheck(sa_session, users, engine):
             print('+%s' % (nice_size(new - current)))
         else:
             print('-%s' % (nice_size(current - new)))
-        if not args.dryrun and engine not in ('postgres', 'postgresql'):
-            user.set_disk_usage(new)
-            sa_session.add(user)
-            sa_session.flush()
 
 
 if __name__ == '__main__':