This replaces N database operations (datasets / 1000), (histories / 1000) with a single SQL query that returns just the number we're interested in.

It's one of the operations that constantly generates logging messages about slow queries, and we force recalculate queries regularly in order to cut down on support emails, so it'd be nice if it was faster.

Does anyone like writing SQLAlchemy? Maybe it's just that my job has become "write plain SQL all day every day" but 90% of the time of this PR was figuring out how to translate the query into SA.