@erasche 2019-08-13
Needs a refactoring so I've moved to WIP, will rewrite quickly.