@erasche 2019-08-13
Ok so here's the deal: Using SQLalchemy with a couple of CTEs, SA decides to "optimise" that into a subquery. The performance of this is 10x worse for me than CTEs.  It works it's just slow :(

If anyone can coerce SA to generate this exact query with the ORM, then, awesome, please fix it :)

Otherwise some numbers that might convince you to just merge my plain sql ;)

User | Datasets | Previous | New
--- | --- | --- | ---
Our Favourite | 2940799 | 29s | 538.730 ms
Bjoern | 48316 | 12s | 180.005 ms


my last python version:

```python
per_user_histories = s.query(History.id) \
    .filter_by(user_id=1, purged=False) \
    .cte("per_user_histories")

per_hist_hdas = s.query(HistoryDatasetAssociation.dataset_id) \
    .distinct(HistoryDatasetAssociation.dataset_id) \
    .filter(HistoryDatasetAssociation.purged == False, HistoryDatasetAssociation.history_id.in_(per_user_histories))\
    .cte("per_hist_hdas")

per_user_datasets = s.query(
        Dataset.id, 
        func.coalesce(Dataset.total_size, func.coalesce(Dataset.file_size)).label("file_size")
    )\
    .outerjoin(LibraryDatasetDatasetAssociation, Dataset.id == LibraryDatasetDatasetAssociation.dataset_id) \
    .filter(LibraryDatasetDatasetAssociation.id == None, Dataset.id.in_(per_hist_hdas))\
    .cte("per_user_datasets")

q = s.query(func.sum(per_user_datasets.c.file_size)).select_from(per_user_datasets)
print(q.all())
print(q.statement)
```