@mvdbeek 2019-08-14
Just checking this one out I noticed that `calculate_disk_usage` AFAICT will only be called if not on postgresql because of https://github.com/erasche/galaxy/blob/7830c54a60369e5faf47ee1c55f2de3604e8251e/lib/galaxy/model/__init__.py#L545

On postgres one would instead call `pgcalc` which does a slightly different query:
https://github.com/erasche/galaxy/blob/7830c54a60369e5faf47ee1c55f2de3604e8251e/lib/galaxy/model/util.py#L6

Could those be united ? I also noticed 100% CPU usage and a hang when attempting to create a new user on a clean new sqlite database, but I haven't checked if that's due to the PR or something being broken on dev.