@mvdbeek 2019-08-14
OK, actually the hang is specific to the PR, it appears on sqlite, probably because of
```
galaxy.queue_worker ERROR 2019-08-14 07:20:41,387 [p:39748,w:4,m:0] [Thread-1] Error running control task type: recalculate_user_disk_usage
Traceback (most recent call last):
  File "/Users/mvandenb/src/galaxy/.venv/lib/python3.7/site-packages/sqlalchemy/engine/base.py", line 1244, in _execute_context
    cursor, statement, parameters, context
  File "/Users/mvandenb/src/galaxy/.venv/lib/python3.7/site-packages/sqlalchemy/engine/default.py", line 550, in do_execute
    cursor.execute(statement, parameters)
sqlite3.OperationalError: wrong number of arguments to function coalesce()

The above exception was the direct cause of the following exception:

Traceback (most recent call last):
  File "lib/galaxy/queue_worker.py", line 349, in process_task
    result = f(self.app, **body['kwargs'])
  File "lib/galaxy/queue_worker.py", line 220, in recalculate_user_disk_usage
    user.calculate_and_set_disk_usage()
  File "lib/galaxy/model/__init__.py", line 548, in calculate_and_set_disk_usage
    new = self.calculate_disk_usage()
  File "lib/galaxy/model/__init__.py", line 533, in calculate_disk_usage
    """, {'user_id': self.id})
  File "/Users/mvandenb/src/galaxy/.venv/lib/python3.7/site-packages/sqlalchemy/orm/session.py", line 1269, in execute
    clause, params or {}
  File "/Users/mvandenb/src/galaxy/.venv/lib/python3.7/site-packages/sqlalchemy/engine/base.py", line 988, in execute
    return meth(self, multiparams, params)
  File "/Users/mvandenb/src/galaxy/.venv/lib/python3.7/site-packages/sqlalchemy/sql/elements.py", line 287, in _execute_on_connection
    return connection._execute_clauseelement(self, multiparams, params)
  File "/Users/mvandenb/src/galaxy/.venv/lib/python3.7/site-packages/sqlalchemy/engine/base.py", line 1107, in _execute_clauseelement
    distilled_params,
  File "/Users/mvandenb/src/galaxy/.venv/lib/python3.7/site-packages/sqlalchemy/engine/base.py", line 1248, in _execute_context
    e, statement, parameters, cursor, context
  File "/Users/mvandenb/src/galaxy/.venv/lib/python3.7/site-packages/sqlalchemy/engine/base.py", line 1466, in _handle_dbapi_exception
    util.raise_from_cause(sqlalchemy_exception, exc_info)
  File "/Users/mvandenb/src/galaxy/.venv/lib/python3.7/site-packages/sqlalchemy/util/compat.py", line 399, in raise_from_cause
    reraise(type(exception), exception, tb=exc_tb, cause=cause)
  File "/Users/mvandenb/src/galaxy/.venv/lib/python3.7/site-packages/sqlalchemy/util/compat.py", line 153, in reraise
    raise value.with_traceback(tb)
  File "/Users/mvandenb/src/galaxy/.venv/lib/python3.7/site-packages/sqlalchemy/engine/base.py", line 1244, in _execute_context
    cursor, statement, parameters, context
  File "/Users/mvandenb/src/galaxy/.venv/lib/python3.7/site-packages/sqlalchemy/engine/default.py", line 550, in do_execute
    cursor.execute(statement, parameters)
sqlalchemy.exc.OperationalError: (sqlite3.OperationalError) wrong number of arguments to function coalesce()
[SQL:
            WITH per_user_histories AS
            (
                SELECT history.id as id
                FROM history
                WHERE history.user_id = ?
                    AND history.purged = false
            ),
            per_hist_hdas AS (
                SELECT DISTINCT history_dataset_association.dataset_id as id
                FROM history_dataset_association
                WHERE history_dataset_association.purged = false
                    AND history_dataset_association.history_id in (SELECT id from per_user_histories)
            )

            SELECT sum(coalesce(dataset.total_size, coalesce(dataset.file_size)))
            FROM dataset
            LEFT OUTER JOIN library_dataset_dataset_association ON dataset.id = library_dataset_dataset_association.dataset_id
            WHERE dataset.id in (SELECT id from per_hist_hdas)
                AND library_dataset_dataset_association.id IS NULL;
        ]
[parameters: (4,)]
(Background on this error at: http://sqlalche.me/e/e3q8)
```