@mvdbeek 2019-08-14
```
diff --git a/lib/galaxy/model/__init__.py b/lib/galaxy/model/__init__.py
index 7193e32ca3..6dfc6a3d11 100644
--- a/lib/galaxy/model/__init__.py
+++ b/lib/galaxy/model/__init__.py
@@ -525,7 +525,7 @@ class User(Dictifiable, RepresentById):
                     AND history_dataset_association.history_id in (SELECT id from per_user_histories)
             )
 
-            SELECT sum(coalesce(dataset.total_size, coalesce(dataset.file_size)))
+            SELECT sum(coalesce(dataset.total_size, dataset.file_size))
             FROM dataset
             LEFT OUTER JOIN library_dataset_dataset_association ON dataset.id = library_dataset_dataset_association.dataset_id
             WHERE dataset.id in (SELECT id from per_hist_hdas)

```
fixes the exception and the hang.